product card click handler updating state of redux store and rendering new related products
