+++ client/src/RelatedItemsComparison/components/ProductCard/ProductCard.jsx
@@ -1,4 +1,5 @@
 import React, { useState } from 'react';
+import { useDispatch } from 'react-redux';
 import PropTypes from 'prop-types';
 import axios from 'axios';
 import styled from 'styled-components';
@@ -20,8 +21,9 @@ const StyledImg = styled.img`
   aspect-ratio: .7;
   object-fit: cover;
 `;
-export default function ProductCard({ id, handleClick }) {
-  // create states for all relevant pieces of data;
+export default function ProductCard({ id }) {
+  const dispatch = useDispatch();
+
   const [photoURL, setPhotoURL] = useState('');
   const [category, setCategory] = useState('');
   const [name, setName] = useState('');
@@ -57,6 +59,18 @@ export default function ProductCard({ id, handleClick }) {
   //   })
   // );
 
+  const handleClick = () => {
+    dispatch({ type: '@product/FETCH_DATA' });
+    axios.get(`/products/${id}`)
+      .then((result) => {
+        console.log('datadata: ', result.data);
+        dispatch({ type: '@product/SET_DATA', payload: result.data });
+      })
+      .catch((err) => {
+        dispatch({ type: '@product/FETCH_FAILED', payload: err.message });
+      });
+  };
+
   // will likely have to wrap this in a useEffect
   axios.all([
     getNameAndCategory(),
@@ -78,7 +92,7 @@ export default function ProductCard({ id, handleClick }) {
     });
 
   return (
-    <Card onClick={() => handleClick(id)}>
+    <Card onClick={() => handleClick()}>
       <StyledImg src={photoURL} />
       <StyledCategory>{category}</StyledCategory>
       <StyledName>{name}</StyledName>
@@ -90,5 +104,4 @@ export default function ProductCard({ id, handleClick }) {
 
 ProductCard.propTypes = {
   id: PropTypes.number.isRequired,
-  handleClick: PropTypes.func.isRequired,
 };
